Development backend

+++ backend/src/services/cardService.js
@@ -1,5 +1,6 @@
 const getNextId = require("../utils/getNextId");
 const repository = require("../repositories/cardRepository");
+const AppError = require("../utils/appError");
 
 async function createCard(data) {
   const id = await getNextId("cardid");
@@ -8,21 +9,36 @@ async function createCard(data) {
     id: id.toString(),
     color: data.color,
     value: data.value,
-    gameId: data.gameId
+    gameId: data.gameId,
   };
 
   return await repository.saveCard(card);
 }
 
 async function getCard(id) {
-  return await repository.findCardById(id);
+  const card = await repository.findCardById(id);
+  if (!card) {
+    throw new AppError("Card not found", 404);
+  }
+
+  return card;
 }
 
 async function updateCard(id, updates) {
+  const card = await repository.findCardById(id);
+  if (!card) {
+    throw new AppError("Card not found", 404);
+  }
+
   return await repository.updateCardById(id, updates);
 }
 
 async function deleteCard(id) {
+  const card = await repository.findCardById(id);
+  if (!card) {
+    throw new AppError("Card not found", 404);
+  }
+
   return await repository.deleteCardById(id);
 }
 
@@ -34,7 +50,7 @@ async function getTopCard(gameId) {
   const card = await repository.getTopDiscardCard(gameId);
 
   if (!card) {
-    throw new Error("No card found for this game");
+    throw new AppError("No card found for this game", 400);
   }
 
   return {
@@ -49,12 +65,11 @@ async function getTopCard(gameId) {
   };
 }
 
-
 module.exports = {
   createCard,
   getCard,
   updateCard,
   deleteCard,
   getAllCards,
-  getTopCard
+  getTopCard,
 };
